Remove crypto3::zk from crypto3::common
