+++ crypto3/libs/marshalling/zk/CMakeLists.txt
@@ -29,9 +29,6 @@ target_link_libraries(${CMAKE_WORKSPACE_NAME}_${CURRENT_PROJECT_NAME} INTERFACE
 
                       crypto3::multiprecision
                       crypto3::algebra
-                      crypto3::zk
-                      crypto3::containers
-                      crypto3::math
 
                       ${CMAKE_WORKSPACE_NAME}::marshalling-multiprecision
                       ${CMAKE_WORKSPACE_NAME}::marshalling-algebra
